fix data not render error
