+++ package.json
@@ -28,6 +28,7 @@
     "jspdf": "^3.0.4",
     "jwt-decode": "^4.0.0",
     "lucide-react": "^0.544.0",
+    "monaco-editor": "^0.55.1",
     "next": "15.5.9",
     "pdf-lib": "^1.17.1",
     "pdf-parse": "^1.1.1",
