+++ src/api/candidateShare/shareCandidate.ts
@@ -1,47 +1,26 @@
 import api from '@/config/axios';
+import { ShareCandidate, ShareCandidatePayload } from '../../types/shareInterfaceCandidate';
 
-interface ShareCandidatePayload {
-  candidateId: string;
-}
-interface Skill {
-  _id?: string;
-  name?: string;
-}
-
-interface Experience {
-  _id?: string;
-  company?: string;
-  role?: string;
-}
-interface User {
-  _id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-}
-
-interface ShareCandidate {
-  _id: string;
-  userId: string;
-  availability: 'looking' | 'not_looking';
-  resumeFile?: string;
-  createdAt: string;
-  updatedAt: string;
-
-  user: User;
-  skills: Skill[];
-  experiences: Experience[];
-}
 export const createShareCandidate = async (data: ShareCandidatePayload[]) => {
   const response = await api.post('/api/share/', {
     users: data.map(item => item.candidateId),
   });
-  console.log('check the response ' + response);
+  console.log(
+    'check the data send from frontend to backend  => ',
+    response,
+    'response.data => ',
+    response.data
+  );
   return response.data;
 };
 
 export const getShareCandidate = async (shareId: string): Promise<ShareCandidate[]> => {
   const response = await api.get(`/api/share/${shareId}`);
-  console.log('check the response of ALL share candidate ===>', response);
+  console.log(
+    'backend response of ALL share candidates ===>',
+    response,
+    'response.data.data ===>',
+    response.data.data
+  );
   return response.data.data;
 };
